requirements/base.txt
Django==3.2
django-environ==0.4.5
django-filter==2.3.0
django-money==2.0
django-phonenumber-field==3.0.1
djangorestframework==3.11.0
phonenumbers==8.10.13
psycopg2==2.9.1
